django_app/requirements.txt
Django==3.2.13
Pillow==9.1.0
djangorestframework==3.12.4
django-storages==1.12.3
boto3==1.9.152